fb: controll user case when user freeze or uninstalls or reinstall apps
till now on navbar navring or powermenu it still shows the app icon
or it does not show any icon after reboot. For UX it is way better
to clearly show the User that she/he has disabled or uninstalled an app
which is binded. If this is the case the Icon changes to the error
none icon....on PowerMenu as well in settings and Error string will be shown
So the user clearly know that he either needs to bind this action to another
one. Or to restore the accidentaly removed app. If the user restores the
Icon of course it shows immediately again up

As well I removed the observer for shortcuts widgets completly and moved
the observing values into the already existend PhoneStatusBar observer.
So we do not have currently on all the systemUI another one then the
actual from AOSP

Change-Id: I25623dc12ba4c1ae6ee6a106aa903d3c28c92a92

Conflicts:
	core/java/com/android/internal/util/slim/AppHelper.java
	core/java/com/android/internal/util/slim/PolicyHelper.java
	packages/SystemUI/src/com/android/systemui/SearchPanelView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/NavigationBarView.java
	packages/SystemUI/src/com/android/systemui/statusbar/phone/PhoneStatusBar.java