[1/2] Base: On-The-Go mode
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

allows to change transparency of the overlay via the notification.

have fun!

Change-Id: I149e3578e51d9b8b8c8bff41c4eaebca774739ab

Conflicts:
	packages/SystemUI/res/values/custom_strings.xml
	policy/src/com/android/internal/policy/impl/GlobalActions.java