Base: On-The-Go: Allow to restart service on the fly
add a toggle to the options dialog to allow the user to restart the service
on the fly, instead of stopping it and create a notification.

it maybe be buggy, depending on how quick their camera can release and
recreate the instance, thats why i went the stop the service way by default

if the user chooses to recreate it on the fly, it resets the views,
triggers a handler with a delay of 1/2 second and sets up the views again

also increate timeout of the dialog, as the transparent dialog may be
harder to read for users and its annoying if the dialog disappears while
you search for the setting you want.

Change-Id: I1233854c10ee524925834c73ea20fba44c0f2a46

Conflicts:
	packages/SystemUI/res/values/custom_strings.xml